INT-2584 - Fix bug where icons get overwritten by codegen -re.

+++ resilient-sdk/resilient_sdk/cmds/codegen.py
@@ -108,6 +108,12 @@ class CmdCodegen(BaseCmd):
             elif isinstance(file_info, str) and os.path.isfile(file_info):
                 # It is just a path to a file, copy it to the target_file
                 target_file = os.path.join(target_dir, file_name)
+                if os.path.exists(target_file):
+                    # If file already exists skip copy.
+                    files_skipped.append(os.path.join(os.path.basename(target_dir), file_name))
+                    continue
+
+                newly_generated_files.append(os.path.join(os.path.basename(target_dir), file_name))
                 shutil.copy(file_info, target_file)
 
             else:
